+++ projects/SelfTest/UsageTests/VariadicMacrosTests.cpp
